Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,24 +2,24 @@
 #include<stdio.h>
 int main(){
     char operator;
-    double num1,num2;
+    int num1,num2;
 
     scanf("%lf %lf", &num1, &num2);
     scanf("%c", &operator);
 
     switch(operator){
         case'+':
-        printf("%.2lf",num1 + num2);
+        printf("%d",num1 + num2);
         break;
         case'-':
-        printf("%.2f", num1 - num2);
+        printf("%d", num1 - num2);
         break;
         case'*':
-        printf("%.2f",num1*num2);
+        printf("%d",num1*num2);
         break;
         case'/':
         if (num2 != 0){
-            printf("%.2f",num1/num2);
+            printf("%d",num1/num2);
         }else{
             printf("Division by zero is not allowed.");
         }
